moved SchemeCache context to cache.h
